Feat: mypage 컴포넌트에서 사용하는 중복 getdocs 재사용 바꿈 by @park74441
